# sprand and SparseMatrixCSC constructor simplifications