This PR implements some simplifications to the sparse code made very easy by #30494. In there, sparse random matrices were effectively constructed from a vector of linear indices (which for `sprand` were taken from `randsubseq`). So I factored out the sparse matrix construction part from `sprand` that can be used now both in `sprand` and in the fallback sparse constructor from `AbstractMatrix` (which does not catch e.g. `Matrix` because there is a more specialized constructor from `StridedMatrix`). I tested that there is no performance regression by checking against `Traspose`. In fact there seems to be a (small) performance gain as (a) there is no need to check for duplicates and (b) using linear indices one avoids the allocation of an extra vector of indices. Is there a way to check for performance regressions more systematically? 
(net outcome from this PR is -60 lines)

```
for n=(10,100,1000), p=(1.0,1e-1,1e-2,1e-3,1e-4)
    @show(n,p)
    Random.seed!(1); x=Transpose(Matrix(sprand(n,n,p))); @show @benchmark sparse($x)
end
```

MASTER:
-------------------------------------------------------------
|    n\p        | 1.0         |   0.1     | 0.01          |  0.001       | 0.0001
|-----------|------------|-----------|---------------|---------------|-------------
| n=10     | 1.481 μs | 517.677 ns | 379.346 ns | 350.789 ns | 353.089 ns
| n=100    | 121.878 μs | 37.005 μs | 20.916 μs | 18.598 μs | 18.476 μs
| n=1000   | 20.618 ms | 6.356 ms | 3.295 ms | 2.892 ms | 2.936 ms
----------------------------------------------------------------------------------


PR:
-------------------------------------------------------------
|    n\p        | 1.0         |   0.1     | 0.01          |  0.001       | 0.0001
|-----------|------------|-----------|---------------|---------------|-------------
| n=10     | 1.357 μs | 445.061 ns | 332.215 ns | 305.885 ns | 307.341 ns
| n=100    | 113.133 μs | 37.731 μs | 20.636 μs | 18.667 μs | 18.342 μs
| n=1000   | 18.671 ms | 6.013 ms | 3.290 ms | 2.924 ms | 2.939 ms
----------------------------------------------------------------------------------