@ViralBShah 2019-01-06
I don't believe that the nanosoldier perf benchmarks have anything relevant here.