@abraunst 2019-01-07
@ViralBShah I am ok with the title change, but note that this is not exclusively about `sprand`. It introduces a contructor of `SparseMatrixCSC` from a linear increasing set of indices, and uses that on `sprand` and `sparse`.