@stevengj 2019-01-07
With this new code, can we get rid of the specialized StridedMatrix constructor, or does the latter still have a substantial performance advantage?