@abraunst 2019-01-07
> With this new code, can we get rid of the specialized StridedMatrix constructor, or does the latter still have a substantial performance advantage?

Unfortunately it is still faster:

```
julia> using SparseArrays, BenchmarkTools

julia> x=Matrix(sprand(1000,1000,0.1));
```

master:

```
julia> @benchmark sparse($x)
BenchmarkTools.Trial: 
  memory estimate:  1.53 MiB
  allocs estimate:  6
  --------------
  minimum time:     1.534 ms (0.00% GC)
  median time:      1.587 ms (0.00% GC)
  mean time:        1.613 ms (1.24% GC)
  maximum time:     27.612 ms (94.07% GC)
  --------------
  samples:          3091
  evals/sample:     1
```

PR:

```
julia> @benchmark sparse($x)
BenchmarkTools.Trial: 
  memory estimate:  4.01 MiB
  allocs estimate:  36
  --------------
  minimum time:     2.333 ms (0.00% GC)
  median time:      2.386 ms (0.00% GC)
  mean time:        2.454 ms (2.57% GC)
  maximum time:     31.432 ms (92.45% GC)
  --------------
  samples:          2036
  evals/sample:     1

```

Noting the extra allocation, I tried the following:

```julia
function SparseMatrixCSC{Tv,Ti}(M::AbstractMatrix) where {Tv,Ti}
    @assert !has_offset_axes(M)
    nz = count(!iszero, M)
    I = Vector{Ti}(undef, nz)
    V = Vector{Tv}(undef, nz)
    i, j = 0, 0

    for v in M
        i += 1
        if !iszero(v)
            I[j += 1] = i
            V[j] = v
        end
    end
    return sparse_sortedlinearindices!(I, V, size(M)...)
end
```

And we get much closer! Will try to benchmark better tomorrow.

```
julia> @benchmark sparse($x)
BenchmarkTools.Trial: 
  memory estimate:  1.54 MiB
  allocs estimate:  6
  --------------
  minimum time:     1.615 ms (0.00% GC)
  median time:      1.664 ms (0.00% GC)
  mean time:        1.703 ms (1.72% GC)
  maximum time:     30.738 ms (94.42% GC)
  --------------
  samples:          2930
  evals/sample:     1
```