@abraunst 2019-01-08
> This is probably fast enough to get rid of the specialized Matrix method. Performance-critical code will probably never use these methods anyway—you always make large sparse matrices directly, not from dense matrices. These conversions are just here for convenience for small interactive problems, I think.

The situation is not clear-cut. Maybe it would be wiser to leave the PR and leave the `M::StridedArrays` constructor?

The above approach is faster for very dense matrices but substantially slower in sparser ones in the `Transpose` test:

master (`AbstractArray` constructor):
```
julia> [ begin Random.seed!(1); x=Transpose(Matrix(sprand(n,n,p)));  b=@benchmark sparse($x); minimum(b.times); end for n=(10,100,1000), p=(1.0,1e-1,1e-2,1e-3,1e-4)]
3×5 Array{Float64,2}:
   1461.6          486.513        366.928        348.009        347.897    
 121178.0        38327.0        20727.0        18619.0        18414.0      
      2.09966e7      6.31002e6      3.36006e6      2.96888e6      2.94113e6
```

above proposal:
```
julia> [ begin Random.seed!(1); x=Transpose(Matrix(sprand(n,n,p)));  b=@benchmark sparse($x); minimum(b.times); end for n=(10,100,1000), p=(1.0,1e-1,1e-2,1e-3,1e-4)]
3×5 Array{Float64,2}:
   470.23         393.356        392.965        382.365        381.788    
 38097.0        40642.0        32357.0        31666.0        31449.0      
     6.36208e6      6.15465e6      5.11586e6      5.11357e6      5.00486e6

```
PR

```
julia> [ begin Random.seed!(1); x=Transpose(Matrix(sprand(n,n,p)));  b=@benchmark sparse($x); minimum(b.times); end for n=(10,100,1000), p=(1.0,1e-1,1e-2,1e-3,1e-4)]
3×5 Array{Float64,2}:
   1262.9          386.683        273.415        236.594        235.64     
 113408.0        38788.0        19654.0        17685.0        17404.0      
      1.70226e7      5.28324e6      3.07331e6      2.79147e6      2.78202e6

```


in the `Matrix` test instead it is the opposite:


master (`M::StridedArray` constructor)
```
julia> [ begin Random.seed!(1); x=Matrix(sprand(n,n,p));  b=@benchmark sparse($x); minimum(b.times); end for n=(10,100,1000), p=(1.0,1e-1,1e-2,1e-3,1e-4)]
3×5 Array{Float64,2}:
   188.217        133.582        128.394     138.25     138.061
 10362.0        11663.0         7455.0      6531.4     6473.6  
     1.68884e6      1.5584e6  799494.0    721780.0   705629.0  

```

above proposal:
```
julia> [ begin Random.seed!(1); x=Matrix(sprand(n,n,p));  b=@benchmark sparse($x); minimum(b.times); end for n=(10,100,1000), p=(1.0,1e-1,1e-2,1e-3,1e-4)]
3×5 Array{Float64,2}:
   246.682        139.053        131.625     132.598     130.24
 15989.0        13112.0         6817.2      5970.83     5806.5 
     2.29853e6      1.6286e6  799516.0    714622.0    696401.0 

```

PR

```
julia> [ begin Random.seed!(1); x=Matrix(sprand(n,n,p));  b=@benchmark sparse($x); minimum(b.times); end for n=(10,100,1000), p=(1.0,1e-1,1e-2,1e-3,1e-4)]
3×5 Array{Float64,2}:
  1112.3          271.748         155.67     120.776     120.347
 95653.0        22313.0          7301.25    5423.17     5190.67 
     9.82539e6      2.33503e6  729518.0   540351.0    520473.0  
```