@abraunst 2019-01-08
The failed check is a timeout:

`The job exceeded the maximum time limit for jobs, and has been terminated.`