@abraunst 2019-01-08
BTW how would people feel about exporting `sparse(I, V, m, n) = sparse_linearindices!(Vector{Int}(undef, length(V)), V, m, n)` ?