@abraunst 2019-01-09
> The situation is not clear-cut. Maybe it would be wiser to leave the PR and leave the `M::StridedArrays` constructor (i.e. PR as is)?

Just to recap: with the PR as-is, we get "PR" for the first set, and "master" for the second set (because we would leave the existent constructor from `StridedArray`. We can choose any combination of the two essentially, but this choice is the only one that is uniformly better than master (in these tests). As you say, this is probably not performance-critical code in any case. Besides the simplification and performance improvement, the PR fixes a bug in  #30627. Let me know how you prefer to proceed!