@KlausC 2019-01-15
> These conversions are just here for convenience for small interactive problems, I think.

The  `SparseMatrixCSC(M::AbstractMatrix)` constructors are also used as the "abstractarray fallback", when you transform for example `Symmetric{Tv,SparseMatrixCSC}`. There is a PR [30552](https://github.com/JuliaLang/julia/pull/30552) under review, which improves the perfomance for those cases essentially.